refactor db calls into db/songReviews
along with some other cleanup
