+++ muze/src/db/songGet.ts
@@ -1,4 +1,4 @@
-import SupabaseClient from './SupabaseClient';;
+import SupabaseClient from '@/db/SupabaseClient';
 
 export default async function getAllSongs() {
   const {data: song} = await SupabaseClient.from("songs").select();
